requirements.txt
absl-py==2.3.1
accelerate==1.12.0
addict==2.4.0
adlfs==2025.8.0
aiobotocore==2.26.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aioitertools==0.13.0
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.13.2
anyio==4.12.1
art==6.5
astunparse==1.6.3
attrs==25.4.0
Authlib==1.6.6
axolotl==0.13.2
axolotl-contribs-lgpl==0.0.7
axolotl-contribs-mit==0.0.6
azure-core==1.38.0
azure-datalake-store==0.0.53
azure-identity==1.25.1
azure-storage-blob==12.28.0
backoff==2.2.1
botocore==1.41.5
brotli==1.2.0
cbor2==5.8.0
certifi==2026.1.4
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
circuitbreaker==2.1.3
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
cryptography==45.0.7
DataProperty==1.1.0
datasets==4.5.0
decorator==5.2.1
dill==0.4.0
distro==1.9.0
einops==0.8.2
evaluate==0.4.1
fastapi==0.128.0
fastcore==1.12.6
ffmpy==1.0.0
filelock==3.20.3
fire==0.7.1
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2025.10.0
gast==0.7.0
gcsfs==2025.10.0
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.48.0rc0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.4
google-cloud-core==2.5.0
google-cloud-storage==3.8.0
google-crc32c==1.8.0
google-generativeai==0.8.6
google-pasta==0.2.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
gradio==6.4.0
gradio_client==2.0.3
groovy==0.1.2
grpcio==1.76.0
grpcio-status==1.71.2
grpclib==0.4.9
h11==0.16.0
h2==4.3.0
h5py==3.15.1
hf-xet==1.2.0
hf_transfer==0.1.9
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
hyperframe==6.1.0
idna==3.11
immutabledict==4.2.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.1.0
joblib==1.5.3
jsonlines==4.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
keras==3.13.1
kernels==0.11.5
langdetect==1.0.9
libclang==18.1.1
llvmlite==0.46.0
lm_eval==0.4.7
lxml==6.0.2
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mbstrdecoder==1.1.4
mdurl==0.1.2
mistral_common==1.8.8
ml_dtypes==0.5.4
modal==1.3.0.post1
more-itertools==10.8.0
mpmath==1.3.0
msal==1.34.0
msal-extensions==1.3.1
multidict==6.7.0
multiprocess==0.70.18
namex==0.1.0
narwhals==2.15.0
networkx==3.6.1
nltk==3.9.2
numba==0.63.1
numexpr==2.14.1
numpy==2.3.5
nvidia-ml-py==12.560.30
oauthlib==3.3.1
oci==2.166.0
ocifs==1.3.2
openenv-core==0.1.0
opt_einsum==3.4.0
optimum==1.16.2
optree==0.18.0
orjson==3.11.5
packaging==26.0
pandas==2.3.3
pathvalidate==3.3.1
peft==0.18.1
pillow==11.3.0
platformdirs==4.5.1
plotly==6.5.2
portalocker==3.2.0
posthog==6.7.11
propcache==0.4.1
proto-plus==1.27.0
protobuf==5.29.5
psutil==7.2.1
pyarrow==23.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pybind11==3.0.1
pycountry==24.6.1
pycparser==3.0
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.1.0
pyparsing==3.3.2
pytablewriter==1.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.22
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
requests-oauthlib==2.0.0
responses==0.18.0
rich==14.2.0
rouge_score==0.1.2
rpds-py==0.30.0
rsa==4.9.1
s3fs==2025.10.0
sacrebleu==2.6.0
safehttpx==0.1.7
safetensors==0.7.0
schedulefree==1.4.1
scikit-learn==1.8.0
scipy==1.17.0
semantic-version==2.10.0
sentencepiece==0.2.1
sentry-sdk==2.50.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sqlitedict==2.1.0
starlette==0.50.0
sympy==1.14.0
synchronicity==0.11.1
tabledata==1.3.4
tabulate==0.9.0
tcolorpy==0.1.7
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.3.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
toml==0.10.2
tomlkit==0.13.3
torch==2.9.1
torchao==0.13.0
tqdm==4.67.1
tqdm-multiprocess==0.0.11
trackio==0.15.0
transformers==4.57.6
trl==0.27.0
typepy==1.3.4
typer==0.21.1
typer-slim==0.21.1
types-certifi==2021.10.8.3
types-toml==0.10.8.20240310
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.40.0
wandb==0.24.0
watchfiles==1.1.1
Werkzeug==3.1.5
word2number==1.1
wrapt==1.17.3
xxhash==3.6.0
yarl==1.22.0
zstandard==0.22.0
